Move all bundles out of the shared dir into main bundles dir
