+++ fronts-client/src/bundles/externalArticlesBundle.ts
