+++ fronts-client/src/shared/actions/__tests__/frontscards.spec.ts
@@ -4,7 +4,7 @@ import thunk from 'redux-thunk';
 import { createArticleEntitiesFromDrop, cardsReceived } from '../Cards';
 import initialState from 'fixtures/initialState';
 import { capiArticle } from '../../fixtures/shared';
-import { actionNames as externalArticleActionNames } from 'shared/bundles/externalArticlesBundle';
+import { actionNames as externalArticleActionNames } from 'bundles/externalArticlesBundle';
 import { createCard } from 'shared/util/card';
 import { RefDrop } from 'util/collectionUtils';
