+++ fronts-client/src/shared/reducers/sharedReducer.ts
@@ -1,7 +1,7 @@
 import cards from './cardsReducer';
 import groups from './groupsReducer';
-import { reducer as collections } from '../bundles/collectionsBundle';
-import { reducer as externalArticles } from '../bundles/externalArticlesBundle';
+import { reducer as collections } from '../../bundles/collectionsBundle';
+import { reducer as externalArticles } from '../../bundles/externalArticlesBundle';
 import { reducer as featureSwitches } from '../redux/modules/featureSwitches';
 import { reducer as pageViewData } from '../../redux/modules/pageViewData';
 import { Card, Group } from 'shared/types/Collection';
