+++ fronts-client/src/shared/selectors/shared.ts
@@ -6,8 +6,8 @@ import {
   getContributorImage,
   isLive
 } from 'util/CAPIUtils';
-import { selectors as externalArticleSelectors } from '../bundles/externalArticlesBundle';
-import { selectors as collectionSelectors } from '../bundles/collectionsBundle';
+import { selectors as externalArticleSelectors } from '../../bundles/externalArticlesBundle';
+import { selectors as collectionSelectors } from '../../bundles/collectionsBundle';
 import { ExternalArticle } from '../types/ExternalArticle';
 import {
   Card,
